requirements.txt
lxml==3.7.1
python-amazon-mws==0.6
python-dateutil==2.6.0
requests==2.12.4
six==1.10.0
wheel==0.24.0
